06/11
Diseño de Reporte Generico cambiado
Reporte Diario y Mensual funcional, diseño de reporte se unifica y se formatea
Reporte Caja falta rediseño de su reporte, correccion de la ventana adaptando al nuevo padron, hay que corregir la logica de la consulta hql para que no envie movimientos anulados
El controlador de la ventana colaborador borra su contenido generando errores, supongo un error del IDE

+++ src/iOS/controlador/util/EventosUtil.java
@@ -180,7 +180,7 @@ public class EventosUtil {
 	public static String separadorMiles(Double valor) {
 		try {
 			DecimalFormat formato = new DecimalFormat("#,###.###;#,###.###");
-			Double valorDouble = ((Double) valor);
+			Double valorDouble = (valor);
 			return formato.format(valorDouble);
 		} catch (Exception e) {
 			return null;
@@ -191,7 +191,7 @@ public class EventosUtil {
 	public static String separadorDecimales(Double valor) {
 		try {
 			DecimalFormat formato = new DecimalFormat("#,##0.0##;#,##0.0##");
-			Double valorDouble = ((Double) valor);
+			Double valorDouble = (valor);
 			return formato.format(valorDouble);
 		} catch (Exception e) {
 			return null;
